refactor: switch to auto generated type definitions
