+++ tsconfig.json
@@ -17,7 +17,9 @@
     "strictFunctionTypes": true,
     "strictNullChecks": true,
     "strictPropertyInitialization": true,
-    "target": "es5"
+    "target": "es5",
+    "declaration": true,
+    "declarationDir": "./dist"
   },
   "files": [
     "src/index.tsx"
