--- types/react-router-breadcrumbs-hoc/index-tests.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import BreadCrumbs from './react-router-breadcrumbs-hoc-tests';
-
-export default function() {
-    return  (
-        <BreadCrumbs />
-    );
-}
